componentize AudioPlayer
